To fix KtlintError, Rename toy_team6_airbnb -> toyTeam6AirBnb (CamelCase)
