+++ src/test/kotlin/com/example/toyTeam6Airbnb/ToyTeam6AirbnbApplicationTests.kt
@@ -1,4 +1,5 @@
-package com.example.toy_team6_airbnb
+
+package com.example.toyTeam6Airbnb
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
@@ -9,5 +10,4 @@ class ToyTeam6AirbnbApplicationTests {
     @Test
     fun contextLoads() {
     }
-
 }
